chore: move constants to scripts/utils
