+++ scripts/utils/migrate/fetch-docs.js
@@ -2,7 +2,7 @@ const fetch = require('node-fetch');
 require('dotenv').config();
 
 const logger = require('../logger');
-const { TYPES, BASE_URL } = require('../../constants');
+const { TYPES, BASE_URL } = require('../constants');
 
 const getUrl = (type) =>
   [BASE_URL, 'api/migration/content', type, 'list'].join('/');
